+++ nodes/iris_call_wrapper.js
@@ -13,8 +13,36 @@ const fs = require("fs");
    
        for (e in subflowJSON['env']){
          let name = subflowJSON['env'][e]['name'];
-         if(name == "Port" || name == "IP" || name == "Namespace" || name == "Username" || name == "Password"){
+         if(name == "Port" || name == "IP" || name == "Namespace" || name == "Username" || name == "Check Connection"){
             subflowJSON['env'][e]['value']  = serverPropsJSON[name];
+         }
+         if(name == "Password" && serverPropsJSON['Hide Password']){
+            let password = {
+               "name": "Password",
+               "type": "cred",
+               "ui": {
+                   "icon": "font-awesome/fa-ellipsis-h"
+               }
+           };
+           subflowJSON['env'][e] = password;
+         }
+         if(name == "Password" && !serverPropsJSON['Hide Password']){
+            let password = {
+               "name": "Password",
+               "type": "str",
+               "value": serverPropsJSON[name],
+               "ui": {
+                   "icon": "font-awesome/fa-ellipsis-h",
+                   "type": "input",
+                   "opts": {
+                       "types": [
+                           "str",
+                           "env"
+                       ]
+                   }
+               }
+           };
+           subflowJSON['env'][e] = password;
          }
        }
     }catch(e){